Reorganize files - pull out styles and pages to own directories
