+++ tenant-file/portal-app/src/App.tsx
@@ -13,7 +13,7 @@ import Dashboard from './component/dashboard';
 import Properties from './component/properties';
 import Home from './component/home';
 import TenantDetails from './component/tenant-details';
-import Tags from './component/Tags-Page';
+import Tags from './pages/Tags-Page';
 
 function App() {
   return (
