+++ tenant-file/portal-app/src/component/tag.tsx
@@ -1,19 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
-import { isDark } from '../utility';
-
-const StyledTag = styled.button`
-  display: inline-block;
-  padding: 0.1rem 0.5rem;
-  background-color: ${(props) => props.color};
-  border-radius: 15px;
-  font-weight: 700;
-  color: ${(props) => {
-    return isDark(props.color!) ? 'white' : 'black';
-  }};
-  box-shadow: none;
-  border: none;
-`;
+import StyledTag from './styles/Tag';
 
 type Tag = {
   name: string;
